+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ## v11.2.9
+* TBC Classic Anniversary support.
 * Fix: #2311 error caused by IconModule_IconEventConditionHandler enabling unconditionally
 
 ## v11.2.8