+++ CHANGELOG.md
@@ -20,6 +20,7 @@ The following items cannot and will not be fixed.
 * Ignore Runes setting for DKs.
 
 ## v11.2.9
+* TBC Classic Anniversary support.
 * Fix: #2311 error caused by IconModule_IconEventConditionHandler enabling unconditionally
 
 ## v11.2.8